refatoração da tela home + vislumbres dos testes
