+++ src/pages/home/index.js
@@ -1,4 +1,7 @@
 import { onNavigate } from '../../utils/history.js';
+import {
+  login, loginGoogle, createuser, currentUser,
+} from '../../services/index.js';
 
 export const Home = () => {
   const home = document.createElement('div');
@@ -33,19 +36,17 @@ export const Home = () => {
   const email = home.querySelector('#email-input');
   const password = home.querySelector('#password-home');
 
-  const userHome = firebase.auth().currentUser;
+  const userHome = currentUser();
 
   // Conectar um usuário com endereço de e-mail e senha
   btn.addEventListener('click', (e) => {
     e.preventDefault();
-    firebase
-      .auth()
-      .signInWithEmailAndPassword(email.value, password.value)
+    login(email.value, password.value)
       .then(() => {
         if (userHome !== null) {
-          onNavigate('/profile');
+          onNavigate('/post');
         } else {
-          onNavigate('/profile');
+          onNavigate('/post');
         }
       })
       .catch((error) => {
@@ -63,23 +64,12 @@ export const Home = () => {
   // Autenticação do Google
   autGoogle.addEventListener('click', (e) => {
     e.preventDefault();
-    const provider = new firebase.auth.GoogleAuthProvider();
-    firebase
-      .auth()
-      .signInWithPopup(provider)
+    loginGoogle()
       .then(() => {
-        const uid = firebase.auth().currentUser.uid;
-        const user = {
-          displayName: firebase.auth().currentUser.displayName,
-          email: firebase.auth().currentUser.email,
-          phoneNumber: firebase.auth().currentUser.phoneNumber,
-          photoUrl: firebase.auth().currentUser.photoURL,
-        };
-        firebase
-          .firestore().collection('users').doc(uid).set({ user });
+        createuser();
       })
       .then(() => {
-        onNavigate('/profile');
+        onNavigate('/post');
       })
       .catch((error) => {
         alert(error.message);
