+++ src/pages/home/index.spec.js
@@ -1,7 +1,11 @@
-import { Home } from "./index.js";
+import { Home } from './index.js';
+// import {currentUser} from '../../services/index.js';
 
-describe("Home", () => {
-  it("should be a function", () => {
-    expect(typeof Home).toBe("function");
+describe('Home', () => {
+  it('should be a function', () => {
+    expect(typeof Home).toBe('function');
+  });
+  it('shoul return homepag', () => {
+    expect(Home()).toMatchSnapshot();
   });
 });
