--- src/services/index.spec.js
@@ -1,8 +1,0 @@
-// importamos la funcion que vamos a testear
-import { myFunction } from "./index";
-
-describe("myFunction", () => {
-  it("should be a function", () => {
-    expect(typeof myFunction).toBe("function");
-  });
-});
