+++ src/style.css
@@ -4,7 +4,8 @@
 /* font-family: 'Roboto', sans-serif; */
 
 
-*Reset do CSS*/ * {
+/*Reset do CSS*/ 
+* {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
